+++ tests/type_inference.py
@@ -75,6 +75,48 @@ def test_as_bool():
     assert_type(unref(result), bool)
 
 
+def test_rx_map():
+    result = Signal(2).rx.map(lambda x: x * 2)
+    assert_type(result, Computed[int])
+    assert_type(unref(result), int)
+
+
+def test_rx_tap():
+    result = Signal(2).rx.tap(lambda x: x + 1)
+    assert_type(result, Computed[int])
+    assert_type(unref(result), int)
+
+
+def test_rx_len():
+    result = Signal([1, 2, 3]).rx.len()
+    assert_type(result, Computed[int])
+    assert_type(unref(result), int)
+
+
+def test_rx_is():
+    result = Signal(10).rx.is_(10)
+    assert_type(result, Computed[bool])
+    assert_type(unref(result), bool)
+
+
+def test_rx_is_not():
+    result = Signal(10).rx.is_not(None)
+    assert_type(result, Computed[bool])
+    assert_type(unref(result), bool)
+
+
+def test_rx_eq():
+    result = Signal(10).rx.eq(10)
+    assert_type(result, Computed[bool])
+    assert_type(unref(result), bool)
+
+
+def test_rx_in():
+    result = Signal("a").rx.in_("cat")
+    assert_type(result, Computed[bool])
+    assert_type(unref(result), bool)
+
+
 def test_str():
     result = str(Signal(1))
     assert_type(result, str)